+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-gui/playground
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ai-gui/react@0.4.1
+  - @ai-gui/core@0.4.1
+  - @ai-gui/vue@0.4.1
+  - @ai-gui/vanilla@0.4.1
+  - @ai-gui/plugin-mermaid@0.4.1
+  - @ai-gui/plugin-citation@0.4.1
+  - @ai-gui/plugin-artifact@0.4.1
+  - @ai-gui/plugin-ui@0.4.1
+  - @ai-gui/plugin-molecule@0.4.1
+  - @ai-gui/plugin-map@0.4.1
+  - @ai-gui/devtools@0.4.1
+
 ## 0.0.2
 
 ### Patch Changes